// src/types/eden/TaxRate.ts
import { t } from 'elysia';
export const TaxRateOne = t.Object({
  SequenceNumber: t.Number(),
  LastModifiedTime: t.Date(),
  TaxCode: t.String(),
  PaidAccount: t.Optional(t.String()),
  RecAccount: t.Optional(t.String()),
  Rate1: t.Number(),
  Date: t.Optional(t.Date()),
  Rate2: t.Optional(t.Number()),
  Combine: t.Optional(t.Number()),
  CombineRate1: t.Optional(t.Number()),
  CombineRate2: t.Optional(t.Number()),
  GSTReceived: t.Optional(t.Number()),
  NetReceived: t.Optional(t.Number()),
  GSTPaid: t.Optional(t.Number()),
  NetPaid: t.Optional(t.Number()),
  RateName: t.Optional(t.String()),
  ReportCycleStart: t.Optional(t.Number()),
  ReportCycleEnd: t.Optional(t.Number()),
  ReportDate: t.Optional(t.Date()),
  PSTReceived: t.Optional(t.Number()),
  PSTPaid: t.Optional(t.Number()),
  Type: t.Optional(t.Number()),
  Combination: t.Optional(t.String()),
  UserNum: t.Optional(t.Number()),
  UserText: t.Optional(t.String()),
  TaggedText: t.Optional(t.String()),
  AliasCode: t.Optional(t.String()),
  AliasCountry: t.Optional(t.String()),
  ReversedRate1: t.Optional(t.Number()),
  ReversedRate2: t.Optional(t.Number()),
});
export const TaxRateMany = t.Object({
  data: t.Array(TaxRateOne),
  pagination: t.Object({
    total: t.Number(),
    limit: t.Number(),
    offset: t.Number(),
    next: t.Number(),
    prev: t.Number(),
  })
});